zadania/Z0203/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
tiktoken==0.8.0